+++ CHANGELOG.md
@@ -9,6 +9,7 @@
   - `Iterator.prototype.drop`
   - `Iterator.prototype.every`
   - `Iterator.prototype.filter`
+  - `Iterator.prototype.find`
 - Compat data improvements:
   - [Explicit Resource Management](https://github.com/tc39/proposal-explicit-resource-management) features disabled (again) in V8 ~ Chromium 135 and re-added in 136
   - [`RegExp.escape`](https://github.com/tc39/proposal-regex-escaping) marked as [shipped from V8 ~ Chromium 136](https://issues.chromium.org/issues/353856236#comment17)